Various cleanup and tweaks now that all the pieces are in
