+++ src/main/java/org/spongepowered/api/util/command/CommandCallable.java
@@ -26,6 +26,7 @@ package org.spongepowered.api.util.command;
 
 import com.google.common.base.Optional;
 import org.spongepowered.api.text.Text;
+import org.spongepowered.api.util.command.spec.CommandSpec;
 
 import java.util.List;
