+++ src/main/java/org/spongepowered/api/util/command/spec/TranslationPlaceholder.java
@@ -1,0 +1,65 @@
+/*
+ * This file is part of SpongeAPI, licensed under the MIT License (MIT).
+ *
+ * Copyright (c) SpongePowered <https://www.spongepowered.org>
+ * Copyright (c) contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package org.spongepowered.api.util.command.spec;
+
+import org.spongepowered.api.text.Text;
+import org.spongepowered.api.text.Texts;
+import org.spongepowered.api.text.translation.Translation;
+
+/**
+ * This class contains methods intended for use as placeholders until true localization is implemented for SpongeAPI strings.
+ */
+class TranslationPlaceholder {
+    private TranslationPlaceholder() {}
+
+    public static Text t(String untranslated, Object... args) {
+        return Texts.of(new FixedTranslation(untranslated), args);
+    }
+
+    private static class FixedTranslation implements Translation {
+        private final String untranslated;
+
+        private FixedTranslation(String untranslated) {
+            this.untranslated = untranslated;
+        }
+
+        @Override
+        public String getId() {
+            return this.untranslated;
+        }
+
+        @Override
+        public String get() {
+            return this.untranslated;
+        }
+
+        @Override
+        public String get(Object... args) {
+            return String.format(this.untranslated, args);
+        }
+    }
+
+
+}
